Added a confirmation box for deleting, changed the background for all pages, and fixed some layout issues with the video player

+++ client/src/views/Admin-Dashboard/ClusterBar.js
@@ -18,37 +18,46 @@ class ClusterBar extends React.Component {
 
 	async deleteCluster() {
 		
-		const clusterid = this.props.cluster.id
+		var answer = window.confirm("Are you sure you want to delete this career?")
+		
+		if(answer === true){
+			const clusterid = this.props.cluster.id
 
-		await axios.delete('http://localhost:5000/api/clusters/deleteCluster/'.concat(clusterid))
-		.then(res => {
-			console.log(res.data)
-			//window.location = res.data.redirect
-		});
-
-		var correctId = 0
-
-		for (var i = 0; i < this.props.clusterslength; i++){
-			
-			if (i != clusterid){
-				
-				const id = {
-					new_id: correctId
-				}
-
-				await axios.put('http://localhost:5000/api/clusters/updateClusterId/'.concat(i), id)
-				.then(res => {
+			await axios.delete('http://localhost:5000/api/clusters/deleteCluster/'.concat(clusterid))
+			.then(res => {
 				console.log(res.data)
-				});
+				//window.location = res.data.redirect
+			});
+	
+			var correctId = 0
+	
+			for (var i = 0; i < this.props.clusterslength; i++){
+				
+				if (i != clusterid){
+					
+					const id = {
+						new_id: correctId
+					}
+	
+					await axios.put('http://localhost:5000/api/clusters/updateClusterId/'.concat(i), id)
+					.then(res => {
+					console.log(res.data)
+					});
+	
+					correctId = correctId + 1;
+				}
+				
+				if (i == this.props.clusterslength - 1) {
+					window.location = "/Admin-Dashboard"
+				}
+			}
 
-				correctId = correctId + 1;
-			}
-			
-			if (i == this.props.clusterslength - 1) {
-				window.location = "/Admin-Dashboard"
-			}
 		}
-
+		
+		else{
+				
+		}
+	
 
 	}
